+++ react-native/storyly-react-native/index.d.ts
@@ -189,6 +189,7 @@ declare module "storyly-react-native" {
       | "text"
       | "image"
       | "producttag"
+      | "productcard"
       | "comment"
       | "video"
       | "vod";