+++ react-native/storyly-react-native/RELEASENOTES.md
@@ -1,4 +1,10 @@
 # Release Notes
+### 1.33.2 (26.04.2023)
+* improved StorylyView initalization flow on native sides
+
+### 1.33.1 (17.04.2023)
+* added logs to ios bridge side
+
 ### 1.33.0 (03.04.2023)
 * updated react dependencies to 18.x.x
 * added prop-types as dependency